tools: Move sandboxing code to separate file
This will help with reuse.
